:recycle: chore: 收敛类型
