+++ src/SortableList/index.ts
@@ -5,4 +5,4 @@ export type { SortableListRef } from './container/StoreUpdater';
 export { useSortableList } from './hooks/useSortableList';
 export type { SortableListInstance } from './hooks/useSortableList';
 export type { SortableListDispatchPayload } from './store';
-export type { CreatorButtonProps } from './type';
+export type { CreatorButtonProps } from './type/type';
